+++ src/lib/accessibility/Behaviors/List/BasicListBehavior.ts
@@ -1,0 +1,11 @@
+import { IAccessibilityDefinition } from '../../interfaces'
+
+const BasicListBehavior: IAccessibilityDefinition = {
+  attributes: {
+    root: {
+      role: 'list',
+    },
+  },
+}
+
+export default BasicListBehavior