+++ src/lib/accessibility/Behaviors/List/BasicListItemBehavior.ts
@@ -1,0 +1,11 @@
+import { IAccessibilityDefinition } from '../../interfaces'
+
+const BasicListItemBehavior: IAccessibilityDefinition = {
+  attributes: {
+    root: {
+      role: 'listitem',
+    },
+  },
+}
+
+export default BasicListItemBehavior