+++ test/specs/components/List/ListItem-test.ts
@@ -1,7 +1,8 @@
-import { isConformant } from 'test/specs/commonTests'
+import { isConformant, handlesAccessibility } from 'test/specs/commonTests'
 
 import ListItem from 'src/components/List/ListItem'
 
 describe('ListItem', () => {
   isConformant(ListItem)
+  handlesAccessibility(ListItem, { defaultRootRole: 'listitem' })
 })